requirements.txt
google-genai==1.2.0
watchdog<=6.0.0
streamlit==1.42.0
# google-cloud-aiplatform[preview]
# google-cloud-secret-manager==2.21.1
gptrim==0.1.6